Establishing JDBC connection in the Filter
You  can create a Connection object connect to JDBC in Servlet to handle
work with Database. However, you can also create a Connection object
connect to JDBC in Filter, and it will take effect with many Servlets.
And you can use this Connection throughout the way of request.  You can
see the illustration below for understandable